feat: add q10 in local test
