+++ src/runtime/src/datasource/nexmark/queries/mod.rs
@@ -13,6 +13,7 @@
 
 mod q0;
 mod q1;
+mod q10;
 mod q13;
 mod q2;
 mod q3;
